Rename useIsMultiplayer to useCollaborationStatus
